+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - '-eval' now evaluates also label names with a starting "_".
 - Dropped support for label files. The labels are extracted anyway from the list file.
 - Watches allow now more complex expressions not only labels. Now e.g. "dlistw+3" is supported.
+- Changed appearance of memory viewer: Added a column header and a legend for the register colors.
 
 
 ## 0.6.2